embedded, gui: fix NCO shift UB, calibration restore, CDCM6208 parse crash
- lms7002m_get_nco_frequency shifted a uint16_t SPI read left by 16; the
  value promotes to signed int, so an MSB-set word (>= 0x8000) was signed
  overflow UB. Cast to uint32_t before the shift.
- A Tx calibration error path called lms7002m_spi_read(self, x0020val),
  reading a register at the data value instead of restoring the channel.
  It now writes x0020val back to register 0x0020, matching the sibling
  path a few lines below.
- CDCM6208 GUI only caught std::invalid_argument from stoi/stod, so an
  out-of-range numeric input threw std::out_of_range and crashed the app.
  It now catches std::logic_error, covering both.

+++ embedded/lms7002m/lms7002m.c
@@ -1246,7 +1246,9 @@ uint32_t lms7002m_get_nco_frequency(lms7002m_context* self, bool isTx, const uin
     const uint32_t refClk_Hz = lms7002m_get_reference_clock_tsp(self, isTx);
     const uint16_t addr = isTx ? 0x0240 : 0x0440;
     uint32_t fcw = 0;
-    fcw |= lms7002m_spi_read(self, addr + 2 + index * 2) << 16; //NCO frequency control word register MSB part.
+    // cast before the shift, a uint16_t promotes to signed int and << 16 of a
+    // value >= 0x8000 would be signed overflow
+    fcw |= (uint32_t)lms7002m_spi_read(self, addr + 2 + index * 2) << 16; //NCO frequency control word register MSB part.
     fcw |= lms7002m_spi_read(self, addr + 3 + index * 2); //NCO frequency control word register LSB part.
 
     return nco_fcw_to_freq(fcw, refClk_Hz);
